requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
aiohttp==3.8.5
urllib3==1.26.16
protobuf==4.24.4
google==3.0.0